+++ src/pages/_app.js
@@ -4,7 +4,7 @@ import Head from 'next/head'
 import { ThemeProvider } from '@material-ui/core/styles'
 import CssBaseline from '@material-ui/core/CssBaseline'
 import theme from '../themes'
-import { useSmoothScroll } from '../Util/SmoothScroll'
+import { useSmoothScroll } from '../utils/SmoothScroll'
 
 export default function MyApp(props) {
   const { Component, pageProps } = props
